ArrMult_4bit Imp & Testbenches
I primarily sourced lab 1 to create these files related to the Array 4-bit multiplier. Note: I initially wanted to use an 8-bit input for the multiplier for consistency so that it matches the input from the ALU; however, we MUST truncate the ALU 8-bit input to the 4-bit multiplier to 4-bits because at any point if you want to calculate the result from the 4-bit multiplier, it will by default/design always create a 2n-bit result relative to the n-bit input size from an n-bit multiplier. This is because to represent the largest value possible from two 4-bit values that you are multiplying, you must double the output to 8-bit.