Time Multiplexer
Here is the time multiplexer with a simple testbench. Works as intended as I already tested it on-board and with a simple behavioral testbench. Note that the inputs must be a 4-bit value for each individual number you want to represent, and it will be translated into a decimal value on-board without overlapping cathodes as a time multiplexer should do.